mainapp/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
groq==0.18.0
numpy==2.2.2
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pytesseract==0.3.13
python-dotenv==1.0.1
reportlab==4.2.5
